+++ src/hictk/CMakeLists.txt
@@ -17,11 +17,17 @@ target_sources(
   hictk_tools
   PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/main.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/cli/cli.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/cli/cli_convert.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/cli/cli_dump.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/cli/cli_load.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/cli/cli_merge.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/cli/cli_zoomify.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/convert/convert.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/convert/cool_to_hic.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/convert/hic_to_cool.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/dump/dump.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/load/load.cpp)
+          ${CMAKE_CURRENT_SOURCE_DIR}/load/load.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/zoomify/zoomify.cpp)
 
 target_include_directories(hictk_tools PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/include)
 target_link_libraries(
