+++ src/hictk/include/hictk/tools/config.hpp
@@ -85,12 +85,23 @@ struct MergeConfig {
   std::uint8_t verbosity{2};
 };
 
+struct ZoomifyConfig {
+  std::string input_uri{};
+  std::string output_path{};
+
+  std::vector<std::uint32_t> resolutions{};
+
+  bool force{false};
+  std::uint8_t verbosity{2};
+};
+
 // clang-format off
 using Config = std::variant<std::monostate,
                             ConvertConfig,
                             DumpConfig,
                             LoadConfig,
-                            MergeConfig>;
+                            MergeConfig,
+                            ZoomifyConfig>;
 // clang-format on
 
 }  // namespace hictk::tools
